+++ project/Settings.scala
@@ -28,7 +28,8 @@ object Settings {
     "-encoding", "utf8",
     "-target:jvm-1.8",
     "-language:postfixOps",
-    "-Ypartial-unification"
+    "-Ypartial-unification",
+    "-language:higherKinds"
   )
 
   //sbt assembly settings
@@ -55,6 +56,5 @@ object Settings {
     //the version is applied in rootVersionSettings and is set to 0.1-githash.
     //we don't really use it for anything but we might when we publish our model
   ) ++ commonAssemblySettings ++ rootVersionSettings
-
 
 }
